refactor(personality): move templates to .md files via include_str!
Each personality template now lives as its own .md file under
crates/zeroclaw-runtime/src/agent/personality_templates/, embedded
via include_str! and substituted with TemplateContext values at
render time. AGENTS.md ships in two variants (regular and
no-memory) since it's the only file whose body changes based on
whether persistent memory is enabled.

Trade-off: Rust source drops from 321 lines to 186; the templates
themselves move into 283 lines of .md files that GitHub renders
properly, editors syntax-highlight, and contributors can hand-edit
without escaping newlines. Binary size is unchanged — same strings,
just embedded from a different source.

Placeholders ({agent}, {user}, {tz}, {comm_style}) render harmlessly
as plain text when the .md files are previewed without rendering.
