+++ crates/zeroclaw-runtime/src/agent/personality_templates/HEARTBEAT.md
@@ -1,0 +1,9 @@
+# HEARTBEAT.md
+
+# Keep this file empty (or with only comments) to skip heartbeat work.
+# Add tasks below when you want {agent} to check something periodically.
+#
+# Examples:
+# - Check my email for important messages
+# - Review my calendar for upcoming events
+# - Run `git status` on my active projects
