+++ crates/zeroclaw-runtime/src/agent/personality_templates/MEMORY.md
@@ -1,0 +1,27 @@
+# MEMORY.md — Long-Term Memory
+
+*Your curated memories. The distilled essence, not raw logs.*
+
+## How This Works
+- Daily files (`memory/YYYY-MM-DD.md`) capture raw events (on-demand via tools)
+- This file captures what's WORTH KEEPING long-term
+- This file is auto-injected into your system prompt each session
+- Keep it concise — every character here costs tokens
+
+## Security
+- ONLY loaded in main session (direct chat with your human)
+- NEVER loaded in group chats or shared contexts
+
+---
+
+## Key Facts
+(Add important facts about your human here)
+
+## Decisions & Preferences
+(Record decisions and preferences here)
+
+## Lessons Learned
+(Document mistakes and insights here)
+
+## Open Loops
+(Track unfinished tasks and follow-ups here)
