+++ crates/zeroclaw-runtime/src/agent/personality_templates/USER.md
@@ -1,0 +1,20 @@
+# USER.md — Who You're Helping
+
+*{agent} reads this file every session to understand you.*
+
+## About You
+- **Name:** {user}
+- **Timezone:** {tz}
+- **Languages:** English
+
+## Communication Style
+- {comm_style}
+
+## Preferences
+- (Add your preferences here — e.g. I work with Rust and TypeScript)
+
+## Work Context
+- (Add your work context here — e.g. building a SaaS product)
+
+---
+*Update this anytime. The more {agent} knows, the better it helps.*
